Time: 9 ms (46.91%), Space: 77.6 MB (9.18%) - LeetHub

+++ 0848-shifting-letters/0848-shifting-letters.java
@@ -10,7 +10,7 @@ class Solution {
         StringBuilder sb = new StringBuilder();
         for(int i=0;i<n;i++){
             int pos = (int) s.charAt(i) - 97;
-            char ch = (char) (((pos + arr[i]) % 26) + 97);
+            char ch = (char) (((long)(pos + arr[i]) % 26) + 97);
 
             sb.append(ch);
         }
